new Perimeter tries to keep track of Family's perimeter
But it is still not working correctly. In part because when
the Family perimeter changes a lot, Elements density will
vary greatly since it is just a function of time.

Maybe the Perimeter should  be the one owning the DistortedCircle
and be used to decide if, when and where Elements are cleated.