Introduces Sockets and NewElementEvent
Sockets have a matrix of places where Elements can be created
and keeps the state of if they've been created. Perimeter uses
these Sockets to know if, when and where new Elements should be
created. Perimeter will throw the event NewElementEvent which
is then caught by Family to create a new Element.