Family now no longer using an FBO
The Surface still uses an FBO but now the Family writes directly
to the Surface. Before it was writing to its own FBO, te Surface
would always clear the screen and then call the Family.draw() which
would simply draw a pre-rendered FBO.

The advantage of using an FBO at Family level is that I can manipulate
each Family independently at any moment. Like removing its color
saturation for example. This can make sense if I want a family to
erode somehow.

The disadvantage of using an FBO is that overimposing the multiple
families was still a problem without a simple solution because of
the transparencies. This is a well known problem with a not that simple
low level OpenGL solution which I never really manage to achieve.

So for now I have simplified the code: no more FBOs at Family level.
If needed I can go back to having them. But for now, the simpler
the better.