Solve GL_BLEND issue
The blending of the hypha with (0,0,0,1) would not go beyond 127
with ofEnableAlphaBlending() on. After A LOT of investigation and
experiments I finally understood that I have to use:
glBlendFunc(GL_ONE, GL_ONE)
so that the colors are properly added. It was stalled for a long
time because I forgot to invoke:
glEnable(GL_BLEND)
and then, at the end:
glDisable(GL_BLEND)
For all this to work it is important that the FBO is initialized
with (0,0,0,0). So now I must do ofClear(settings.backgroundColor)
instead of using settings.backgroundColor to initialize the FBO.