Correct bug in Border
It was rounding to int. I found that the offset could not be too big
and this is probably because it was defined as int. I defined it as
float and the problem went away.

I also got rid of the ANGLE_RESOLUTION. Now Border only has 360
values which are more than enough.