Rewrote Border so that it keeps track of which points were touched
And, once touched, it returns true for that particular angle regarless
of the position. Added Border.isOutside() and removed
Hypha.isOutsideBorder() because this really belonged inside Border.
With this new Border algorithm, Hyphae no longer kills all Hypha
once the first one touches the border, allowing it to continue to
approach a shape which is much more similar with the Border.