Replace GPU instance drawing with drawing an ofVboMesh
This was suggested by Arturo here:
https://forum.openframeworks.cc/t/shader-not-drawing-points-1-to-1-why/31924

While I was able to draw 280k with the instance drawing, I can still
draw 240k which is quite ok considering how simpler this approach
is and considering that it solves the problem I was having with
anti-aliasing and scaling. This is now drawing 1-to-1 as required.