Add ofDisableAntiAliasing() to solve the, well, antialiasing
Now the GPU version is about 1.5x faster than the CPU version and
it looks exactly the same!
And I believe that there is something wrong with macOS Mojave and,
if there weren't, this would be even faster. The reason I say there
is something wrong is because the example vboExample is supposed to
show that ofVboMesh is much faster than ofMesh and yet, in my computer,
there is no difference. And I tried it in another mac with High Sierra
and the difference was noticeable.