Stop sending ofVec2f direction to Hypha and send float angle instead
Also no need to store a direction. This was done so that an angle offset
could be applied to a primal hypha and propagated to its children but,
as a bonus, it simplified the code and it probably made it faster.

While doing this I realised that the check for maxBentAngle didn't
make sense and so I removed it. The setting is still there and needs
another implementation. That or remove it altogether.